# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/organizations/admin/kyc/pending": {
      "get": {
        "operationId": "OrganizationsController_getPendingKycSubmissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/api/v1/kyc/case/{caseId}": {
      "get": {
        "operationId": "KycController_getKYCCase",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KYC"
        ]
      }
    },
    "/api/v1/catalog": {
      "post": {
        "operationId": "CatalogController_createCatalog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create my catalog (Seller only)",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/me": {
      "get": {
        "operationId": "CatalogController_getMyCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my catalog",
        "tags": [
          "Catalog"
        ]
      },
      "put": {
        "operationId": "CatalogController_updateMyCatalog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update my catalog",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/me/categories": {
      "post": {
        "operationId": "CatalogController_addCategory",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add category to catalog",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/me/grades": {
      "post": {
        "operationId": "CatalogController_addGrade",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add grade to catalog",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/org/{orgId}": {
      "get": {
        "operationId": "CatalogController_getCatalogByOrgId",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get catalog by organization ID (Public)",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/search": {
      "get": {
        "operationId": "CatalogController_searchCatalogs",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grade",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search catalogs by category/grade",
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/settlements/batches": {
      "post": {
        "operationId": "SettlementsController_createBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create settlement batch (Admin only)",
        "tags": [
          "Settlements"
        ]
      },
      "get": {
        "operationId": "SettlementsController_getAllBatches",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all settlement batches (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
...
    "/api/v1/settlements/batches/{id}/run-payouts": {
      "post": {
        "operationId": "SettlementsController_runPayouts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPayoutsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run payouts for batch (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
...
    "/api/v1/buyer/orders": {
      "get": {
        "operationId": "BuyerController_getMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Buyer"
        ]
      }
    },
    "/api/v1/buyer/payments": {
      "get": {
        "operationId": "BuyerController_getMyPayments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Buyer"
        ]
      }
    },
...
    "/api/v1/admin/kyc/pending": {
      "get": {
        "operationId": "KycAdminController_getPendingKycCases",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
    "/api/v1/admin/kyc/case/{kycCaseId}": {
      "get": {
        "operationId": "KycAdminController_getKycCaseById",
        "parameters": [
          {
            "name": "kycCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
    "/api/v1/admin/kyc/case/{kycCaseId}/approve": {
      "post": {
        "operationId": "KycAdminController_approveKyc",
        "parameters": [
          {
            "name": "kycCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
    "/api/v1/admin/kyc/case/{kycCaseId}/reject": {
      "post": {
        "operationId": "KycAdminController_rejectKyc",
        "parameters": [
          {
            "name": "kycCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
    "/api/v1/admin/kyc/org/{orgId}/history": {
      "get": {
        "operationId": "KycAdminController_getKycHistory",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
...
    "/api/v1/admin/settlements/batches": {
      "post": {
        "operationId": "SettlementsAdminController_createBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      },
      "get": {
        "operationId": "SettlementsAdminController_listBatches",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      }
    },
    "/api/v1/admin/settlements/batches/{id}/run-payouts": {
      "post": {
        "operationId": "SettlementsAdminController_runPayouts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPayoutsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      }
    },
...
  },
  "info": {
    "title": "B2B Marketplace API",
    "description": "Backend API for B2B Marketplace Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUYER",
              "SELLER",
              "ADMIN",
              "3PL"
            ]
          },
          "organizationName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "mobile",
          "role",
          "organizationName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateCatalogDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCatalogDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBatchDto": {
        "type": "object",
        "properties": {}
      },
      "RunPayoutsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}